+++ packages/sdk/src/codegen/templates/typescript/hooks.template.ts
@@ -15,13 +15,12 @@ import type {
 	SubscriptionHook,
 	SubscriptionHookWithoutInput,
 } from "@wundergraph/sdk/server";
-import type { InternalClient } from "./wundergraph.internal.client"
 import type { User } from "./wundergraph.server"
 import { InternalOperationsClient } from "./wundergraph.internal.operations.client";
 
 export type DATA_SOURCES = {{{dataSourcesUnion}}};
 										
-export interface HookContext<TCustomContext = any> extends BaseRequestContext<User, InternalClient, InternalOperationsClient, TCustomContext> {}
+export interface HookContext<TCustomContext = any> extends BaseRequestContext<User, InternalOperationsClient, TCustomContext> {}
 
 export type HooksConfig<TCustomContext = any> = HooksConfiguration<
 	QueryHooks<TCustomContext>,
@@ -62,7 +61,7 @@ export interface UploadHooks<TCustomContext = any> {
             {{#each uploadProfiles}}
                 {{@key}}?: {
                     preUpload?: (hook: PreUploadHookRequest<User, TCustomContext>) => PreUploadHookResponse;
-                    postUpload?: (hook: PostUploadHookRequest<User, InternalClient, TCustomContext>) => PostUploadHookResponse;
+                    postUpload?: (hook: PostUploadHookRequest<User, TCustomContext>) => PostUploadHookResponse;
                 }
             {{/each}}
         }