+++ packages/sdk/src/webhooks/factory.ts
@@ -1,39 +1,32 @@
 import { OperationsClient } from '../server/operations-client';
-import { InternalClient, Webhook, WebhookHttpEvent, WebhookHttpResponse, WebhookRequestContext } from '../server';
+import { Webhook, WebhookHttpEvent, WebhookHttpResponse, WebhookRequestContext } from '../server';
 import type { ORM } from '../orm';
 
 export const createWebhookFactory = <
 	InternalOperations extends OperationsClient = OperationsClient,
-	TInternalClient extends InternalClient = InternalClient,
 	TCustomContext = any,
 	TypedORM = ORM<any>
 >() => {
-	return createWebhook<InternalOperations, TInternalClient, TCustomContext, TypedORM>();
+	return createWebhook<InternalOperations, TCustomContext, TypedORM>();
 };
 
 export interface CreateWebhookProps<
 	TEvent,
 	TResponse,
 	TOperationsClient extends OperationsClient,
-	TInternalClient extends InternalClient,
 	TCustomContext = any,
 	TypedORM = ORM<any>
 > {
 	handler: (
 		event: TEvent,
-		context: WebhookRequestContext<TInternalClient, TOperationsClient, TCustomContext, TypedORM>
+		context: WebhookRequestContext<TOperationsClient, TCustomContext, TypedORM>
 	) => Promise<TResponse>;
 }
 
-const createWebhook = <
-	TOperationsClient extends OperationsClient,
-	TInternalClient extends InternalClient,
-	TCustomContext = any,
-	TypedORM = ORM<any>
->() => {
+const createWebhook = <TOperationsClient extends OperationsClient, TCustomContext = any, TypedORM = ORM<any>>() => {
 	return <TEvent extends WebhookHttpEvent, TResponse extends WebhookHttpResponse>(
-		webhook: CreateWebhookProps<TEvent, TResponse, TOperationsClient, TInternalClient, TCustomContext, TypedORM>
-	): Webhook<TInternalClient, TEvent, TResponse, TOperationsClient> => {
+		webhook: CreateWebhookProps<TEvent, TResponse, TOperationsClient, TCustomContext, TypedORM>
+	): Webhook<TEvent, TResponse, TOperationsClient> => {
 		return webhook;
 	};
 };