+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.14.1...@wundergraph/nextjs@0.14.2) (2023-09-05)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
 ## [0.14.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.14.0...@wundergraph/nextjs@0.14.1) (2023-09-01)
 
 **Note:** Version bump only for package @wundergraph/nextjs